directory tree refactored, started upgrading weapon logic. All simple weapons implemented.

+++ src/main/java/item/Item.java
@@ -7,5 +7,5 @@ public abstract class Item implements Proficiency {
     protected String name;
     protected Coin cost;
     /** in lb */
-    protected int weight;
+    protected double weight;
 }
